redesigned query process, fixed bugs
