+++ front/src/query.js
@@ -1,0 +1,13 @@
+var query = {
+    ALL_SPORTSMEN: 'sportsman/all',
+    ALL_ACCOMODATIONS: 'accomodation/all',
+    ALL_VOLUNTEERS: 'volunteers/all',
+
+    GET_SPORTSMEN: 'sportsman/get',
+    GET_ACCOMODATIONS: 'accomodation/get',
+    GET_VOLUNTEERS: 'volunteers/get',
+
+    SET_SPORTSMEN: 'sportsman/set'
+};
+
+module.exports = query;
